initial refactor

+++ src/main/scala/krs/User.scala
@@ -2,6 +2,6 @@ package krs
 
 case class User(
   name: String,
-  creditScore: Short,
+  creditScore: Long,
   outstandingLoanAmount: Double
 )
